back/requirements.txt
Flask
flask-cors
mysql-connector-python
gunicorn
flask-socketio